Add planrepotrigger model and signal handler for build_complete
This commit adds the PlanRepositoryTrigger model, and an associated
signal handler for build_complete that will call its `fire()` method.
This method enqueues a build for the instance's listed plan_repo.

TODO: Do we need a validation that prevents a PRT's repo/plan_repo from
pointing to the same repository?