Fix a heisenbug (more info in commit msg)
The old code picked a random item from Plan.objects.all() at module initialization time.

This would have been a record from the real database, not the (mostly empty)
test db. If it happened to randomly pick a low number, the mistake would
be benign. But if it picked a high number it would refer to an object
that doesn't exist at test time.

For some reason this was only detected by Pytest during test cleanup.