Make it possible to store Robot Framework screenshots as non-public S3 objects
This changes the code which imports robot test result XML to convert screenshot URLs into an intermediate representation like asset://1234 (using the id of the TestResultAsset in the database)
instead of directly into a public S3 URL. And it changes the view which renders those results to convert the intermediate URL into an S3 URL that includes time-limited query string authorization.

This is meant for use with the new S3 buckets we're moving to.
Existing test results still keep their old screenshot URLs, which will stop working once the old S3 buckets are deleted.