Support 'Skip' as a valid status for robot tests
We were setting the status to "Pass" if robot status is "PASS", or "Fail"
otherwise. However, robot 4.x now supports a status of "SKIP" in addition to
"PASS" or "FAIL", so it should be safe to just change the case to sync
up with one of the metaci values "Pass", "Fail", or "Skip".

Tragically, "SKIP" could mean that the test didn't run, or that it ran
and failed but was marked as SKIP in a teardown or via the "--skiponfailure"
command line option. I really don't like the way they implemented this. :-(.